Move ensure_admin_roles to internal_role_service.py
