+++ src/layman/upgrade/upgrade_v1_23.py
@@ -2,7 +2,7 @@ import logging
 
 from db import util as db_util
 from layman import settings
-from layman.authz import role_service as role_service_util
+from layman.authz import internal_role_service
 
 logger = logging.getLogger(__name__)
 DB_SCHEMA = settings.LAYMAN_PRIME_SCHEMA
@@ -69,7 +69,7 @@ from {DB_SCHEMA}.users u inner join
 ;"""
     db_util.run_statement(create_layman_users_user_roles_view)
 
-    role_service_util.ensure_admin_roles()
+    internal_role_service.ensure_admin_roles()
 
     create_roles_view = f"""create view {ROLE_SERVICE_SCHEMA}.roles
 as
