+++ src/layman/upgrade/upgrade_v1_23_test.py
@@ -2,7 +2,7 @@ import pytest
 
 from db import util as db_util
 from layman import app, settings
-from layman.authz import role_service as role_service_util
+from layman.authz import internal_role_service
 from layman.common.prime_db_schema import ensure_whole_user
 from test_tools import process_client
 from . import upgrade_v1_23
@@ -123,7 +123,7 @@ def test_create_role_service_schema():
         result = db_util.run_query(user_roles_query)[0]
         assert result[0] + result[1] + result[2] == result[3]
 
-        role_service_util.ensure_admin_roles()
+        internal_role_service.ensure_admin_roles()
         result = db_util.run_query(roles_query)[0]
         assert result[0] + result[1] + result[2] == result[3]
         result = db_util.run_query(user_roles_query)[0]
